+++ R/expect_doppelganger.R
@@ -3,5 +3,5 @@
 #' @noRd
 expect_doppelganger <- function(title, fig, path = NULL, ...) {
   testthat::skip_if_not_installed("vdiffr")
-  vdiffr::expect_doppelganger(title, fig, path = path, ...)
+  vdiffr::expect_doppelganger(title, fig, ...)
 }