+++ R/theme_genes.R
@@ -25,13 +25,13 @@
 theme_genes <- function() {
   ggplot2::theme_grey() + ggplot2::theme(
     panel.background = ggplot2::element_blank(),
-    panel.grid.major.y = ggplot2::element_line(colour = "grey", size = 1),
+    panel.grid.major.y = ggplot2::element_line(colour = "grey", linewidth = 1),
     panel.grid.minor.y = ggplot2::element_blank(),
     panel.grid.minor.x = ggplot2::element_blank(),
     panel.grid.major.x = ggplot2::element_blank(),
     axis.ticks.y = ggplot2::element_blank(),
-    axis.line.x = ggplot2::element_line(colour = "grey20", size = 0.5),
-    axis.ticks.x = ggplot2::element_line(colour = "grey20", size = 0.5),
+    axis.line.x = ggplot2::element_line(colour = "grey20", linewidth = 0.5),
+    axis.ticks.x = ggplot2::element_line(colour = "grey20", linewidth = 0.5),
     strip.text = ggplot2::element_blank(),
     strip.background = ggplot2::element_blank()
   )
@@ -42,13 +42,13 @@ theme_genes <- function() {
 theme_genes_flipped <- function() {
   ggplot2::theme_grey() + ggplot2::theme(
     panel.background = ggplot2::element_blank(),
-    panel.grid.major.x = ggplot2::element_line(colour = "grey", size = 1),
+    panel.grid.major.x = ggplot2::element_line(colour = "grey", linewidth = 1),
     panel.grid.minor.x = ggplot2::element_blank(),
     panel.grid.minor.y = ggplot2::element_blank(),
     panel.grid.major.y = ggplot2::element_blank(),
     axis.ticks.x = ggplot2::element_blank(),
-    axis.line.y = ggplot2::element_line(colour = "grey20", size = 0.5),
-    axis.ticks.y = ggplot2::element_line(colour = "grey20", size = 0.5),
+    axis.line.y = ggplot2::element_line(colour = "grey20", linewidth = 0.5),
+    axis.ticks.y = ggplot2::element_line(colour = "grey20", linewidth = 0.5),
     strip.text = ggplot2::element_blank(),
     strip.background = ggplot2::element_blank()
   )