+++ tests/testthat/test-geom_feature.R
@@ -1,7 +1,7 @@
 context("geom_feature")
 
 test_that("a simple plot with features is drawn without errors", {
-  expect_silent( {
+  expect_no_error( {
     library(ggplot2)
     p <- ggplot(example_genes, aes(xmin = start, xmax = end, y = molecule, fill = gene)) +
       geom_feature(data = example_features, aes(x = position, y = molecule,