+++ tests/testthat/test-geom_gene_label.R
@@ -1,7 +1,7 @@
 context("geom_gene_label")
 
 test_that("geom_gene_label can be added to a plot without error", {
-  expect_silent( {
+  expect_no_error( {
   p <- ggplot2::ggplot(
     example_genes,
     ggplot2::aes(xmin = start, xmax = end, y = molecule, label = gene)